Move ambex snapshot-dumping code out of main.go into snapshots.go.
