+++ pkg/ambex/main.go
@@ -40,6 +40,7 @@ package ambex
 
 import (
 	// standard library
+
 	"context"
 	"encoding/json"
 	"flag"
@@ -303,111 +304,6 @@ func Decode(ctx context.Context, name string) (proto.Message, error) {
 	}
 	dlog.Infof(ctx, "Loaded file %s", name)
 	return v, nil
-}
-
-// Observability:
-//
-// These "expanded" snapshots make the snapshots we log easier to read: basically,
-// instead of just indexing by Golang types, make the JSON marshal with real names.
-type v2ExpandedSnapshot struct {
-	Endpoints ecp_v2_cache.Resources `json:"endpoints"`
-	Clusters  ecp_v2_cache.Resources `json:"clusters"`
-	Routes    ecp_v2_cache.Resources `json:"routes"`
-	Listeners ecp_v2_cache.Resources `json:"listeners"`
-	Runtimes  ecp_v2_cache.Resources `json:"runtimes"`
-}
-
-func NewV2ExpandedSnapshot(v2snap *ecp_v2_cache.Snapshot) v2ExpandedSnapshot {
-	return v2ExpandedSnapshot{
-		Endpoints: v2snap.Resources[ecp_cache_types.Endpoint],
-		Clusters:  v2snap.Resources[ecp_cache_types.Cluster],
-		Routes:    v2snap.Resources[ecp_cache_types.Route],
-		Listeners: v2snap.Resources[ecp_cache_types.Listener],
-		Runtimes:  v2snap.Resources[ecp_cache_types.Runtime],
-	}
-}
-
-type v3ExpandedSnapshot struct {
-	Endpoints ecp_v3_cache.Resources `json:"endpoints"`
-	Clusters  ecp_v3_cache.Resources `json:"clusters"`
-	Routes    ecp_v3_cache.Resources `json:"routes"`
-	Listeners ecp_v3_cache.Resources `json:"listeners"`
-	Runtimes  ecp_v3_cache.Resources `json:"runtimes"`
-}
-
-func NewV3ExpandedSnapshot(v3snap *ecp_v3_cache.Snapshot) v3ExpandedSnapshot {
-	return v3ExpandedSnapshot{
-		Endpoints: v3snap.Resources[ecp_cache_types.Endpoint],
-		Clusters:  v3snap.Resources[ecp_cache_types.Cluster],
-		Routes:    v3snap.Resources[ecp_cache_types.Route],
-		Listeners: v3snap.Resources[ecp_cache_types.Listener],
-		Runtimes:  v3snap.Resources[ecp_cache_types.Runtime],
-	}
-}
-
-// A combinedSnapshot has both a V2 and V3 snapshot, for logging.
-type combinedSnapshot struct {
-	Version string             `json:"version"`
-	V2      v2ExpandedSnapshot `json:"v2"`
-	V3      v3ExpandedSnapshot `json:"v3"`
-}
-
-// csDump creates a combinedSnapshot from a V2 snapshot and a V3 snapshot, then
-// dumps the combinedSnapshot to disk. Only numsnaps snapshots are kept: ambex-1.json
-// is the newest, then ambex-2.json, etc., so ambex-$numsnaps.json is the oldest.
-// Every time we write a new one, we rename all the older ones, ditching the oldest
-// after we've written numsnaps snapshots.
-func csDump(ctx context.Context, snapdirPath string, numsnaps int, generation int, v2snap *ecp_v2_cache.Snapshot, v3snap *ecp_v3_cache.Snapshot) {
-	if numsnaps <= 0 {
-		// Don't do snapshotting at all.
-		return
-	}
-
-	// OK, they want snapshots. Make a proper version string...
-	version := fmt.Sprintf("v%d", generation)
-
-	// ...and a combinedSnapshot.
-	cs := combinedSnapshot{
-		Version: version,
-		V2:      NewV2ExpandedSnapshot(v2snap),
-		V3:      NewV3ExpandedSnapshot(v3snap),
-	}
-
-	// Next up, marshal as JSON and write to ambex-0.json. Note that we
-	// didn't say anything about a -0 file; that's because it's about to
-	// be renamed.
-
-	bs, err := json.MarshalIndent(cs, "", "  ")
-
-	if err != nil {
-		dlog.Errorf(ctx, "CSNAP: marshal failure: %s", err)
-		return
-	}
-
-	csPath := path.Join(snapdirPath, "ambex-0.json")
-
-	err = ioutil.WriteFile(csPath, bs, 0644)
-
-	if err != nil {
-		dlog.Errorf(ctx, "CSNAP: write failure: %s", err)
-	} else {
-		dlog.Infof(ctx, "Saved snapshot %s", version)
-	}
-
-	// Rotate everything one file down. This includes renaming the just-written
-	// ambex-0 to ambex-1.
-	for i := numsnaps; i > 0; i-- {
-		previous := i - 1
-
-		fromPath := path.Join(snapdirPath, fmt.Sprintf("ambex-%d.json", previous))
-		toPath := path.Join(snapdirPath, fmt.Sprintf("ambex-%d.json", i))
-
-		err := os.Rename(fromPath, toPath)
-
-		if (err != nil) && !os.IsNotExist(err) {
-			dlog.Infof(ctx, "CSNAP: could not rename %s -> %s: %#v", fromPath, toPath, err)
-		}
-	}
 }
 
 // Get an updated snapshot going.
